requirements.txt
monai
torch
tensorboardX
neptune-client
numpy